+++ src/extension.ts
@@ -3,8 +3,8 @@ import { IBMiTestManager } from "./manager";
 import { getComponentRegistry, getInstance, loadBase } from "./extensions/ibmi";
 import { Configuration, Section } from "./configuration";
 import IBMi from "@halcyontech/vscode-ibmi-types/api/IBMi";
-import { RPGUnit } from "./components/rpgUnit";
-import { CodeCov } from "./components/codeCov";
+import { RPGUnit } from "./components/rpgUnit/rpgUnit";
+import { CodeCov } from "./components/codeCov/codeCov";
 import * as tmp from "tmp";
 import { TestOutputLogger } from "./loggers/testOutputLogger";
 import { TestStubGenerator } from "./codeActions/testStubGenerator";
